refactor: removed the duplicat Skeleton component
